requirements.txt
streamlit
gspread
google-auth-oauthlib
pandas
altair
requests